+++ src/Timesheet/Day.java
@@ -5,7 +5,7 @@ import java.time.LocalTime;
 public interface Day {
   //  long[] minutesByTypes = new Long[];
     void removeSlot(Slot slot);
-  long totalWorkedMinutes();
+  double totalWorkedMinutes();
   void addTimeslot(LocalTime from, LocalTime until);
   void addBreakSlot(LocalTime from,LocalTime until);
 }
